+++ src/types/api/index.ts
@@ -7,3 +7,4 @@ export * from './auth';
 export * from './forms';
 export * from './language';
 export * from './page';
+export * from './system';